pyproject.toml
[project]
name = "autoresearcher-litter-detection"
version = "0.1.0"
description = "Autonomous CNN training for litter segmentation"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "mlflow>=2.13.0",
    "datasets>=2.19.0",
    "Pillow>=10.3.0",
    "numpy>=1.26.0",
    "matplotlib>=3.9.0",
    "tqdm>=4.66.0",
    "scikit-learn>=1.5.0",
    "pycocotools>=2.0.8",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "ipywidgets>=8.1.0",
    "huggingface-hub>=0.23.0",
    "albumentations>=1.4.0",
    "opencv-python-headless>=4.10.0",
    "eclipse-zenoh>=1.9.0",
    "opencv-python>=4.13.0.92",
    "opentelemetry-sdk>=1.24",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24",
    "opentelemetry-semantic-conventions>=0.45",
    "python-dotenv>=1.2.2",
]
[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }